ecg-ai/requirements.txt
fastapi
uvicorn
tensorflow==2.15.0
numpy
python-dotenv
openai
pydantic